Fix failing unit test
The unit tests failed with the following error:
IOError: [Errno 2] No such file or directory: 'data/example_dump.xml'

This was because the example data file 'example_dump.xml' was hard-referenced.
This was prone to errors during execution (though it did not occur locally).

Change that to infer the position relatively to the test file
(using OS methods dirname with the magic word __file__).