Delay printing of CSV lines in cat2cohort
We want our code to be reusable in different contexts.
The method list_users handles the API querying and result parsing,
which is useful in its own right. There are reasons for having it
not printing the results as CSV format directly:
- it is more reusable this way, since we might want to export
  to other formats.
- it opens the way for more advanced CSV handling.

Transform list_users from a dead-end method to a generator method
yielding couples (username, language).
Move the printing part to cat_to_cohort.