Extract method for CSV line formatting
We need to format couples (user, language) as a WikiMetrics
CSV line, which is "<username>, <wiki>"
where <wiki> is the concatenation of the language and the wiki name
(such as 'frwiki' for fr.wikipedia.
This format could be changed in the future,
and as we only support Wikipedia wikis at the moment in cat2cohort,
we might want to support other wikis as well.
This well warants to have this logic in its own method.

Extract the formatting code to method _make_CSV_line.
Add unittest for new method _make_CSV_line.