Extract method for full CSV formatting
We need to format a userlist (ie, a list of couples (user, language))
as a WikiMetrics compatible cohort, in CSV format.
(Note it uses the method introduced in e4b9823.)

It makes sense to have this logic in a separate method to use it
in different contexts (user lists might be obtained in a different way than

Extract the formatting code to method _userlist_to_CSV_cohort.
Add unittest for new method _userlist_to_CSV_cohort.