Print CSV cohort only when used as a script
We want to be able to use cat2cohort in various contexts.
One of them is in script mode, where we want the CSV cohort to be printed,
but we might want to use it as a library and the CSV must be processed.

Change cat_to_cohort to return the CSV rather than printing it.
Move the print statemetn in main method.