Add class DumpMediaCollection
We add a class DumpMediaCollection, which inherits from dict,
and serve as wrapper around the dump parsed from the XML.

This will allow to manipulate the collection more easily,
by adding methods or implementing caching for example.

Add method init_from_xml_dump which updates the media collection
with the contents parsed from a XML dump.

Note that the MediaCollection could be initialized through
another way, like API queries for example, and is not specifically
tied to the concept of an XML dump.