Add method to retrieve the latest revision of a page
We are often interested in the latest revision of a Page,
because it shows its current contents.

Add method to access this latest revision. It trivially returns
the last revision of the list of revisions, assuming that the list
is ordered - which is the case so far with the XML initialisation.

But this abstraction method also makes sense because we might want
to add other ways to retrieve the latest revision (eg through the API).