Get the state of a collection at any given time
"As a program leader with an interest in media collections,
I want to be able to restore the state of a
given media collection at any point of time
in order to perform retroactive analysis."

This implements the above user-story.
The method get_state takes as parameter
a date (as a datetime object) and returns a
MediaCollection in the state it was at the time.
This means it only contains the revisions before
the target date, and therefore the files not yet
uploaded at the time are not part of it.

As we get back a MediaCollection, we can then use
all the power of this object for analysis.

Closes #2.