Get the initial state of a collection.
The initial state is the collection of all the files
at their first revision. It is useful to make diffs with it,
in order to measure the editing activty on Wikimedia Commons.

Note that since state methods return MediaCollections,
it is possible to compose them, to get the initial state of
a collection as it existed at some point in time
(before subsequent uploads).