Get the differential between two dates
As a program leader, I want to know how much activity
happened in a collection between two given dates.

Add a method which retrieves the number of revisions,
of edits, and of editors on the collection between two dates.

Add companion method, which calls the above and writes
a human readable text report of it.

This completes our feature set from the old Trutat prototype.