App: Dirty fix for Category prefix bug
The underlying object behind FDC reporting, Commons_cat_metrics,
takes as parameter a category name without the prefix.

But our shiny autocomplete adds the prefix, and my JS-fu has not
yet solved this (though it is probably not hard).

In the meantime, quick'n'dirty Python hack to make sure
the prefix is removed if it is there.