App: Convert HTML templates to use inheritance
We defined in aa2296d a base Jinja template to harmonise
and factorise the layout of the different HTML pages.

Here all the HTML templates are changed to inherit
from the base template.

(In the process we changed the lyout of the result pages,
which may not be a good idea)