App: Make Flask generate the links in the navigation bar
The links to the root and the tools in the navbar used to
be static links to a relative root (like "./fdc")

These links break if used in a URL subfolder, which is the case
in the results pages of the tools (like '/wm-metrics/fdc/submit').

Flask provides a way to generate URLs based on the actual method
in the Python file, which solves the problem.