Remove dependency of CommonsCatMetrics on FDC.Round
CommonsCatMetrics has no reason to be tied to the concept
of the FDC round − it should merely apply to a time period.

Here we remove the coupling to FDC.Round. It is the jib of the caller
(ie Indicators.py and still the CLI script) to convert the FDC round
to a time period − which is made easy by the method added in 67ca7fd)